+++ src/renderer/helpers/Normalizer.js
